api with notification completed
