+++ Backend/src/model/booking.ts
@@ -18,7 +18,6 @@ const BookingSchema = new mongoose.Schema({
         total_price : {type : Number},
     },
     coupon_id:{type:Number,ref:'coupons'},
-    status:{type:String},
     paymentId:{type:String},
     orderId:{type:String}
 })
